finished collecting usage statistics user prompt
